Fix with python3.6.

+++ sippy/SipCiscoGUID.py
@@ -43,7 +43,7 @@ class SipCiscoGUID(SipGenericHF):
         else:
             salt = str((random() * 1000000000) + time())
             s = md5(salt.encode()).hexdigest()
-            self.ciscoGUID = (long(s[0:8], 16), long(s[8:16], 16), long(s[16:24], 16), long(s[24:32], 16))
+            self.ciscoGUID = (int(s[0:8], 16), int(s[8:16], 16), int(s[16:24], 16), int(s[24:32], 16))
 
     def parse(self):
         self.ciscoGUID = tuple([int(x) for x in  self.body.split('-', 3)])
